+++ jdaviz/configs/default/plugins/collapse/collapse.py
@@ -1,9 +1,12 @@
+import os
+from pathlib import Path
 import warnings
 
+from astropy.nddata import CCDData
 from glue.core import Data
 from specutils import Spectrum1D
 from specutils.manipulation import spectral_slab
-from traitlets import List, Unicode, observe
+from traitlets import Bool, List, Unicode, observe
 
 from jdaviz.core.events import SnackbarMessage
 from jdaviz.core.registries import tray_registry
@@ -39,12 +42,21 @@ class Collapse(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixi
     template_file = __file__, "collapse.vue"
     function_items = List().tag(sync=True)
     function_selected = Unicode('Sum').tag(sync=True)
+    filename = Unicode().tag(sync=True)
+    collapsed_spec_available = Bool(False).tag(sync=True)
+    overwrite_warn = Bool(False).tag(sync=True)
+    # export_enabled controls whether saving to a file is enabled via the UI.  This
+    # is a temporary measure to allow server-installations to disable saving server-side until
+    # saving client-side is supported
+    export_enabled = Bool(True).tag(sync=True)
 
     def __init__(self, *args, **kwargs):
 
         super().__init__(*args, **kwargs)
 
         self._label_counter = 0
+
+        self.collapsed_spec = None
 
         self.function = SelectPluginComponent(self,
                                               items='function_items',
@@ -53,6 +65,11 @@ class Collapse(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixi
 
         self.dataset.add_filter('is_cube')
         self.add_results.viewer.filters = ['is_image_viewer']
+
+        if self.app.state.settings.get('server_is_remote', False):
+            # when the server is remote, saving the file in python would save on the server, not
+            # on the user's machine, so export support in cubeviz should be disabled
+            self.export_enabled = False
 
     @property
     def _default_spectrum_viewer_reference_name(self):
@@ -93,6 +110,12 @@ class Collapse(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixi
             # Spatial-spatial image only.
             collapsed_spec = spec.collapse(self.function_selected.lower(), axis=-1).T  # Quantity
 
+            # stuff for exporting to file
+            self.collapsed_spec = collapsed_spec
+            self.collapsed_spec_available = True
+            fname_label = self.dataset_selected.replace("[", "_").replace("]", "")
+            self.filename = f"collapsed_{self.function_selected.lower()}_{fname_label}.fits"
+
         if add_data:
             data = Data()
             data['flux'] = collapsed_spec.value
@@ -110,3 +133,46 @@ class Collapse(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixi
 
     def vue_collapse(self, *args, **kwargs):
         self.collapse(add_data=True)
+
+    def vue_save_as_fits(self, *args):
+        self._save_collapsed_spec_to_fits()
+
+    def vue_overwrite_fits(self, *args):
+        """Attempt to force writing the file if the user confirms the desire to overwrite."""
+        self.overwrite_warn = False
+        self._save_collapsed_spec_to_fits(overwrite=True)
+
+    def _save_collapsed_spec_to_fits(self, overwrite=False, *args):
+
+        if not self.export_enabled:
+            # this should never be triggered since this is intended for UI-disabling and the
+            # UI section is hidden, but would prevent any JS-hacking
+            raise ValueError("Writing out collapsed cube to file is currently disabled")
+
+        # Make sure file does not end up in weird places in standalone mode.
+        path = os.path.dirname(self.filename)
+        if path and not os.path.exists(path):
+            raise ValueError(f"Invalid path={path}")
+        elif (not path or path.startswith("..")) and os.environ.get("JDAVIZ_START_DIR", ""):  # noqa: E501 # pragma: no cover
+            filename = Path(os.environ["JDAVIZ_START_DIR"]) / self.filename
+        else:
+            filename = Path(self.filename).resolve()
+
+        if filename.exists():
+            if overwrite:
+                # Try to delete the file
+                filename.unlink()
+                if filename.exists():
+                    # Warn the user if the file still exists
+                    raise FileExistsError(f"Unable to delete {filename}. Check user permissions.")
+            else:
+                self.overwrite_warn = True
+                return
+
+        filename = str(filename)
+        CCDData(self.collapsed_spec).write(filename)
+
+        # Let the user know where we saved the file.
+        self.hub.broadcast(SnackbarMessage(
+            f"Collapsed cube saved to {os.path.abspath(filename)}",
+                           sender=self, color="success"))